requirements.txt
openai-whisper
selenium